ajout makefile et correction

+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: flmuller <flmuller@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/06 15:59:24 by flmuller          #+#    #+#             */
-/*   Updated: 2023/10/08 19:11:43 by flmuller         ###   ########.fr       */
+/*   Updated: 2023/10/09 11:48:46 by flmuller         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,13 +18,13 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	char	*str;
 
 	i = 0;
-	while (s[start + i] && i < len)
+	while (s[start + i] && i < (int) len)
 		i++;
 	str = malloc(sizeof(char) * i + 1);
 	if (!str)
 		return (NULL);
 	i = 0;
-	while (s[start + i] && i < len)
+	while (s[start + i] && i < (int) len)
 	{
 		str[i] = s[start + i];
 		i++;
